Request Response models added, Response Handler Added. Network failure hadling Done, Database Sync Model completed, Background service to send data to server added with volley integration.

+++ app/src/main/java/com/android/prasadmukne/datadrivenmechanic/commons/database/SQLiteDatabaseManager.java
@@ -112,7 +112,7 @@ public class SQLiteDatabaseManager
 		return sqLiteDatabase.delete(tableName, WhereClause, selectionArgs)> 0;
 	}
 
-	public boolean deleteOnRowId(String tableName, long rowId)
+	public boolean deleteOnRowId(String tableName, int rowId)
 	{
 		return sqLiteDatabase.delete(tableName, "_id" + "=" + rowId, null) > 0;
 	}
